+++ examples/vue/sign-up-in/src/ContentPage.vue
@@ -4,11 +4,10 @@ import { KeyringProviderInjectionKey } from '@w3ui/vue-keyring'
 
 export default {
   inject: {
+    account: { from: KeyringProviderInjectionKey.account },
     agent: { from: KeyringProviderInjectionKey.agent },
-    space: { from: KeyringProviderInjectionKey.space },
-    createSpace: { from: KeyringProviderInjectionKey.createSpace },
-    registerSpace: { from: KeyringProviderInjectionKey.registerSpace },
-    cancelRegisterSpace: { from: KeyringProviderInjectionKey.cancelRegisterSpace },
+    authorize: { from: KeyringProviderInjectionKey.authorize },
+    cancelAuthorize: { from: KeyringProviderInjectionKey.cancelAuthorize },
     unloadAgent: { from: KeyringProviderInjectionKey.unloadAgent }
   },
   setup: function(){
@@ -23,21 +22,20 @@ export default {
   computed: {
   },
   methods: {
-    async handleRegisterSubmit (e) {
+    async handleAuthorizeSubmit (e) {
       e.preventDefault()
       this.submitted = true
       try {
-        await this.createSpace()
-        await this.registerSpace(this.email)
+        await this.authorize(this.email)
       } catch (err) {
-        throw new Error('failed to register', { cause: err })
+        throw new Error('failed to authorize', { cause: err })
       } finally {
         this.submitted = false
       }
     },
-    handleCancelRegisterSubmit (e) {
+    handleCancelAuthorizeSubmit (e) {
       e.preventDefault()
-      this.cancelRegisterSpace()
+      this.cancelAuthorize()
     },
     handleSignOutSubmit (e) {
       e.preventDefault()
@@ -48,8 +46,8 @@ export default {
 </script>
 
 <template>
-  <div v-if="space?.registered()">
-    <h1 className="near-white">Welcome {{agent.email}}!</h1>
+  <div v-if="account">
+    <h1 className="near-white">Welcome {{account}}!</h1>
     <p>You are logged in!!</p>
     <form @submit="handleSignOutSubmit">
       <button type="submit" className="ph3 pv2">Sign Out</button>
@@ -58,17 +56,17 @@ export default {
 
   <div v-if="submitted">
     <h1 className="near-white">Verify your email address!</h1>
-    <p>Click the link in the email we sent to {{agent.email}} to sign in.</p>
-    <form @submit="handleCancelRegisterSubmit">
+    <p>Click the link in the email we sent to {{account}} to sign in.</p>
+    <form @submit="handleCancelAuthorizeSubmit">
       <button type="submit" className="ph3 pv2">Cancel</button>
     </form>
   </div>
 
-  <form v-if="!space?.registered() && !submitted" @submit="handleRegisterSubmit">
+  <form v-if="!account && !submitted" @submit="handleAuthorizeSubmit">
     <div className="mb3">
       <label htmlFor="email" className="db mb2">Email address:</label>
       <input id="email" className="db pa2 w-100" type="email" v-model="email" required />
     </div>
-    <button type="submit" className="ph3 pv2" :disabled="submitted">Register</button>
+    <button type="submit" className="ph3 pv2" :disabled="submitted">Authorize</button>
   </form>
 </template>