+++ packages/react-keyring/src/providers/Keyring.tsx
@@ -1,21 +1,11 @@
-import React, { createContext, useState, useContext } from 'react'
-import useLocalStorageState from 'use-local-storage-state'
-import {
-  createAgent,
-  Space,
-  getCurrentSpace as getCurrentSpaceInAgent,
-  getSpaces,
-  CreateDelegationOptions
-} from '@w3ui/keyring-core'
 import type {
+  Agent,
+  Abilities,
   KeyringContextState,
   KeyringContextActions,
   ServiceConfig,
   RegisterSpaceOpts
 } from '@w3ui/keyring-core'
-import type { Agent } from '@web3-storage/access'
-import type { Abilities } from '@web3-storage/access/types'
-import { authorizeWithSocket } from '@web3-storage/access/agent'
 import type {
   Capability,
   Delegation,
@@ -24,6 +14,17 @@ import type {
   Proof,
   Signer
 } from '@ucanto/interface'
+
+import React, { createContext, useState, useContext } from 'react'
+import useLocalStorageState from 'use-local-storage-state'
+import {
+  authorize as accessAuthorize,
+  createAgent,
+  Space,
+  getCurrentSpace as getCurrentSpaceInAgent,
+  getSpaces,
+  CreateDelegationOptions
+} from '@w3ui/keyring-core'
 
 export { KeyringContextState, KeyringContextActions }
 
@@ -49,7 +50,7 @@ export const keyringContextDefaultValue: KeyringContextValue = [
     },
     setCurrentSpace: async () => {},
     registerSpace: async () => {},
-    cancelRegisterSpace: () => {},
+    cancelAuthorize: () => {},
     getProofs: async () => [],
     createDelegation: async () => {
       throw new Error('missing keyring context provider')
@@ -94,13 +95,13 @@ export function KeyringProvider ({
     return agent
   }
 
-  const authorize = async (email: '{string}@{string}'): Promise<void> => {
+  const authorize = async (email: `${string}@${string}`): Promise<void> => {
     const agent = await getAgent()
     const controller = new AbortController()
     setRegisterAbortController(controller)
 
     try {
-      await authorizeWithSocket(agent, email, { signal: controller.signal })
+      await accessAuthorize(agent, email, { signal: controller.signal })
       // TODO is there other state that needs to be initialized?
       setAccount(email)
       const newSpaces = getSpaces(agent)
@@ -116,7 +117,7 @@ export function KeyringProvider ({
     }
   }
 
-  const cancelRegisterSpace = (): void => {
+  const cancelAuthorize = (): void => {
     if (registerAbortController != null) {
       registerAbortController.abort()
     }
@@ -132,21 +133,11 @@ export function KeyringProvider ({
 
   const registerSpace = async (email: string, opts: RegisterSpaceOpts = {}): Promise<void> => {
     const agent = await getAgent()
-    const controller = new AbortController()
-    setRegisterAbortController(controller)
-
-    try {
-      await agent.registerSpace(email, {
-        signal: controller.signal,
-        provider: opts.provider ?? (agent.connection.id.did() as DID<'web'>)
-      })
-      setSpace(getCurrentSpaceInAgent(agent))
-      setSpaces(getSpaces(agent))
-    } catch (error) {
-      if (!controller.signal.aborted) {
-        throw error
-      }
-    }
+    await agent.registerSpace(email, {
+      provider: opts.provider ?? (agent.connection.id.did() as DID<'web'>)
+    })
+    setSpace(getCurrentSpaceInAgent(agent))
+    setSpaces(getSpaces(agent))
   }
 
   const setCurrentSpace = async (did: DID): Promise<void> => {
@@ -165,6 +156,7 @@ export function KeyringProvider ({
     setSpaces([])
     setIssuer(undefined)
     setAgent(undefined)
+    setAccount(undefined)
   }
 
   const resetAgent = async (): Promise<void> => {
@@ -209,12 +201,12 @@ export function KeyringProvider ({
   }
   const actions = {
     authorize,
+    cancelAuthorize,
     loadAgent,
     unloadAgent,
     resetAgent,
     createSpace,
     registerSpace,
-    cancelRegisterSpace,
     setCurrentSpace,
     getProofs,
     createDelegation,