+++ packages/vue-keyring/src/providers/Keyring.ts
@@ -1,3 +1,10 @@
+import type {
+  Agent,
+  KeyringContextState,
+  KeyringContextActions,
+  ServiceConfig
+} from '@w3ui/keyring-core'
+import type { Capability, DID, Proof } from '@ucanto/interface'
 import {
   defineComponent,
   provide,
@@ -6,20 +13,17 @@ import {
   Ref,
   shallowReactive
 } from 'vue'
-import { createAgent, getCurrentSpace as getCurrentSpaceInAgent, getSpaces } from '@w3ui/keyring-core'
-import type {
-  KeyringContextState,
-  KeyringContextActions,
-  ServiceConfig
+import {
+  authorize as accessAuthorize,
+  createAgent,
+  getCurrentSpace as getCurrentSpaceInAgent,
+  getSpaces
 } from '@w3ui/keyring-core'
-import { authorizeWithSocket } from '@web3-storage/access/agent'
-
-import type { Agent } from '@web3-storage/access'
-import type { Capability, DID, Proof } from '@ucanto/interface'
 
 export { KeyringContextState, KeyringContextActions }
 
 interface KeyringProviderInjectionKeyType {
+  account: InjectionKey<Ref<KeyringContextState['account']>>
   space: InjectionKey<Ref<KeyringContextState['space']>>
   spaces: InjectionKey<Ref<KeyringContextState['spaces']>>
   agent: InjectionKey<Ref<KeyringContextState['agent']>>
@@ -29,15 +33,16 @@ interface KeyringProviderInjectionKeyType {
   createSpace: InjectionKey<KeyringContextActions['createSpace']>
   setCurrentSpace: InjectionKey<KeyringContextActions['setCurrentSpace']>
   registerSpace: InjectionKey<KeyringContextActions['registerSpace']>
-  cancelRegisterSpace: InjectionKey<KeyringContextActions['cancelRegisterSpace']>
   getProofs: InjectionKey<KeyringContextActions['getProofs']>
   authorize: InjectionKey<KeyringContextActions['authorize']>
+  cancelAuthorize: InjectionKey<KeyringContextActions['cancelAuthorize']>
 }
 
 /**
  * Injection keys for keyring provider context.
  */
 export const KeyringProviderInjectionKey: KeyringProviderInjectionKeyType = {
+  account: Symbol('w3ui keyring account'),
   space: Symbol('w3ui keyring space'),
   spaces: Symbol('w3ui keyring spaces'),
   agent: Symbol('w3ui keyring agent'),
@@ -47,9 +52,9 @@ export const KeyringProviderInjectionKey: KeyringProviderInjectionKeyType = {
   createSpace: Symbol('w3ui keyring createSpace'),
   setCurrentSpace: Symbol('w3ui keyring setCurrentSpace'),
   registerSpace: Symbol('w3ui keyring registerSpace'),
-  cancelRegisterSpace: Symbol('w3ui keyring cancelRegisterSpace'),
   getProofs: Symbol('w3ui keyring getProofs'),
-  authorize: Symbol('w3ui keyring authorize')
+  authorize: Symbol('w3ui keyring authorize'),
+  cancelAuthorize: Symbol('w3ui keyring cancelAuthorize')
 }
 
 export interface KeyringProviderProps extends ServiceConfig { }
@@ -80,6 +85,10 @@ export const KeyringProvider = defineComponent<KeyringProviderProps>({
       KeyringProviderInjectionKey.spaces,
       computed(() => state.spaces)
     )
+    provide(
+      KeyringProviderInjectionKey.account,
+      computed(() => state.account)
+    )
 
     const getAgent = async (): Promise<Agent> => {
       if (agent == null) {
@@ -92,13 +101,13 @@ export const KeyringProvider = defineComponent<KeyringProviderProps>({
     }
 
     provide(KeyringProviderInjectionKey.authorize,
-      async (email: '{string}@{string}'): Promise<void> => {
+      async (email: `${string}@${string}`): Promise<void> => {
         const agent = await getAgent()
         const controller = new AbortController()
         registerAbortController = controller
 
         try {
-          await authorizeWithSocket(agent, email, { signal: controller.signal })
+          await accessAuthorize(agent, email, { signal: controller.signal })
           // TODO is there other state that needs to be initialized?
           state.account = email
           const newSpaces = getSpaces(agent)
@@ -114,7 +123,7 @@ export const KeyringProvider = defineComponent<KeyringProviderProps>({
         }
       })
 
-    provide(KeyringProviderInjectionKey.cancelRegisterSpace, (): void => {
+    provide(KeyringProviderInjectionKey.cancelAuthorize, (): void => {
       if (registerAbortController != null) {
         registerAbortController.abort()
       }
@@ -169,6 +178,7 @@ export const KeyringProvider = defineComponent<KeyringProviderProps>({
       state.space = undefined
       state.spaces = []
       state.agent = undefined
+      state.account = undefined
       agent = undefined
     }
     provide(KeyringProviderInjectionKey.unloadAgent, unloadAgent)